feat: use marquee everywhere
