+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -   Optimized memory usage and startup time of cable models. After updating, cables will appear disconnected, but this is only visual. Cause a block update to fix this.
 -   Optimized performance of searching in the Grid.
+-   Custom titles that overflow will now have a marquee effect instead, for every GUI.
 
 ### Fixed
